@@ -1,7 +1,7 @@
 {
   "name": "node-red-contrib-uibuilder",
-  "version": "7.1.0",
-  "description": "Easily create data-driven web UI's for Node-RED using any (or no) front-end library.",
+  "version": "7.2.0",
+  "description": "Easily create data-driven web UI's for Node-RED. Single- & Multi-page. Multiple UI's. Work with existing web development workflows or mix and match with no-code/low-code features.",
   "scripts": {
     "preinstall": "node ./bin/uibpreinstalljs",
     "watch": "gulp watch",
@@ -17,15 +17,18 @@
     "express": "^4.17.1",
     "fast-glob": "^3.2.7",
     "fs-extra": "^11.1.0",
-    "jsdom": "^25.0.0",
+    "jsdom": "^26.0.0",
     "socket.io": "^4.3.2"
   },
   "devDependencies": {
+    "@eslint/js": "^9.20.0",
+    "@stylistic/eslint-plugin": "^4.0.1",
     "@types/express": "^4.17.13",
     "@types/jquery": "^3.5.25",
     "@types/jqueryui": "^1.12.19",
     "@types/node": "^18.19.24",
     "@types/node-red": "*",
+    "browserslist": "^4.24.4",
     "docsify": "^4.13.1",
     "docsify-copy-code": "^3.0.0",
     "docsify-darklight-theme": "^3.2.0",
@@ -33,17 +36,14 @@
     "docsify-plugin-flexible-alerts": "^1.1.1",
     "docsify-plugin-toc": "^1.3.2",
     "docsify-themeable": "^0.9.0",
-    "esbuild": "^0.24.0",
-    "eslint": "^8.57.0",
-    "eslint-config-standard": "^17.1.0",
-    "eslint-plugin-es": "^4.1.0",
-    "eslint-plugin-html": "^8.1.1",
-    "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-jsdoc": "^50.2.2",
-    "eslint-plugin-n": "^16.6.2",
-    "eslint-plugin-promise": "^6.6.0",
-    "eslint-plugin-sonarjs": "^2.0.0",
+    "esbuild": "^0.25.0",
+    "eslint": "^9.20.1",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-jsdoc": "^50.6.3",
+    "eslint-plugin-n": "^17.15.1",
+    "eslint-plugin-promise": "^7.2.1",
     "execa": "^5",
+    "globals": "^16.0.0",
     "gulp": "^5.0.0",
     "gulp-clean-css": "^4.3.0",
     "gulp-esbuild": "^0.14.0",
@@ -73,6 +73,7 @@
       "uib-html": "nodes/uib-html/customNode.js",
       "uib-save": "nodes/uib-save/customNode.js",
       "uib-sender": "nodes/uib-sender/uib-sender.js",
+      "uib-sidebar": "nodes/uib-sidebar/customNode.js",
       "uib-tag": "nodes/uib-tag/customNode.js",
       "uib-update": "nodes/uib-update/customNode.js"
     },
