@@ -184,7 +184,7 @@ function detailsPage(uib, urlPrefix) {
         delete uibSummary.me
         delete uibSummary.RED
     } catch (e) {
-        log.warn(`[uibuilder:apiv2:detailsPage] Could not parse uib object, some info will be missing. ${e.message}`)
+        log.warn(`🌐⚠️[uibuilder:apiv2:detailsPage] Could not parse uib object, some info will be missing. ${e.message}`)
     }
 
     // Build the web page
@@ -383,7 +383,6 @@ function detailsPage(uib, urlPrefix) {
 
     // Show the ExpressJS paths currently defined
     // web.routers contains descriptive info on routes, routes.user contains the ExpressJS route stack info.
-    // console.log('>> web >>', web.htmlBuildTable( web.routers.user, ['name', 'desc', 'path', 'type', 'folder'] ))
     page += `
         <h2>uibuilder ExpressJS Routes</h2>
         <p>These tables show all of the web URL routes for uibuilder.</p>
@@ -458,7 +457,7 @@ function adminRouterV2(uib, log) {
         // @ts-ignore
         const chkUrl = chkParamUrl(params)
         if ( chkUrl.status !== 0 ) {
-            log.error(`[uibuilder:apiv2:uibgetfile] Admin API. ${chkUrl.statusMessage}`)
+            log.error(`🌐🛑[uibuilder:apiv2:uibgetfile] Admin API. ${chkUrl.statusMessage}`)
             res.statusMessage = chkUrl.statusMessage
             res.status(chkUrl.status).end()
             return
@@ -467,7 +466,7 @@ function adminRouterV2(uib, log) {
         // @ts-ignore
         const chkFname = chkParamFname(params)
         if ( chkFname.status !== 0 ) {
-            log.error(`[uibuilder:apiv2:uibgetfile] Admin API. ${chkFname.statusMessage}. url=${params.url}`)
+            log.error(`🌐🛑[uibuilder:apiv2:uibgetfile] Admin API. ${chkFname.statusMessage}. url=${params.url}`)
             res.statusMessage = chkFname.statusMessage
             res.status(chkFname.status).end()
             return
@@ -476,14 +475,14 @@ function adminRouterV2(uib, log) {
         // @ts-ignore
         const chkFldr = chkParamFldr(params)
         if ( chkFldr.status !== 0 ) {
-            log.error(`[uibuilder:apiv2:uibgetfile] Admin API. ${chkFldr.statusMessage}. url=${params.url}`)
+            log.error(`🌐🛑[uibuilder:apiv2:uibgetfile] Admin API. ${chkFldr.statusMessage}. url=${params.url}`)
             res.statusMessage = chkFldr.statusMessage
             res.status(chkFldr.status).end()
             return
         }
         //#endregion ---- ----
 
-        log.trace(`[uibuilder:apiv2:uibgetfile] Admin API. File get requested. url=${params.url}, file=${params.folder}/${params.fname}`)
+        log.trace(`🌐[uibuilder[:apiv2:uibgetfile] Admin API. File get requested. url=${params.url}, file=${params.folder}/${params.fname}`)
 
         // if fldr = root, no folder
         if ( params.folder === 'root' ) params.folder = ''
@@ -509,7 +508,7 @@ function adminRouterV2(uib, log) {
                 }
             )
         } else {
-            log.error(`🛑 [uibuilder:apiv2:uibgetfile] Admin API. File does not exist '${filePath}'. url=${params.url}`)
+            log.error(`🌐🛑[uibuilder:apiv2:uibgetfile] Admin API. File does not exist '${filePath}'. url=${params.url}`)
             res.statusMessage = 'File does not exist'
             res.status(500).end()
         }
@@ -526,7 +525,7 @@ function adminRouterV2(uib, log) {
 
         const chkUrl = chkParamUrl(params)
         if ( chkUrl.status !== 0 ) {
-            log.error(`[uibuilder:apiv2:uibputfile] Admin API v2. ${chkUrl.statusMessage}`)
+            log.error(`🌐🛑[uibuilder:apiv2:uibputfile] Admin API v2. ${chkUrl.statusMessage}`)
             res.statusMessage = chkUrl.statusMessage
             res.status(chkUrl.status).end()
             return
@@ -534,7 +533,7 @@ function adminRouterV2(uib, log) {
 
         const chkFname = chkParamFname(params)
         if ( chkFname.status !== 0 ) {
-            log.error(`[uibuilder:apiv2:uibputfile] Admin API. ${chkFname.statusMessage}. url=${params.url}`)
+            log.error(`🌐🛑[uibuilder:apiv2:uibputfile] Admin API. ${chkFname.statusMessage}. url=${params.url}`)
             res.statusMessage = chkFname.statusMessage
             res.status(chkFname.status).end()
             return
@@ -542,14 +541,14 @@ function adminRouterV2(uib, log) {
 
         const chkFldr = chkParamFldr(params)
         if ( chkFldr.status !== 0 ) {
-            log.error(`[uibuilder:apiv2:uibputfile] Admin API. ${chkFldr.statusMessage}. url=${params.url}`)
+            log.error(`🌐🛑[uibuilder:apiv2:uibputfile] Admin API. ${chkFldr.statusMessage}. url=${params.url}`)
             res.statusMessage = chkFldr.statusMessage
             res.status(chkFldr.status).end()
             return
         }
         //#endregion ====== ====== //
 
-        log.trace(`[uibuilder:apiv2:uibputfile] Admin API. File put requested. url=${params.url}, file=${params.folder}/${params.fname}, reload? ${params.reload}`)
+        log.trace(`🌐[uibuilder[:apiv2:uibputfile] Admin API. File put requested. url=${params.url}, file=${params.folder}/${params.fname}, reload? ${params.reload}`)
 
         // Fix for Issue #155 - if fldr = root, no folder
         if ( params.folder === 'root' ) params.folder = '.'
@@ -560,12 +559,12 @@ function adminRouterV2(uib, log) {
         fslib.writeFileCb(fullname, req.body.data, function (err) {
             if (err) {
                 // Send back a response message and code 200 = OK, 500 (Internal Server Error)=Update failed
-                log.error(`🛑 [uibuilder:apiv2:uibputfile] Admin API. File write FAIL. url=${params.url}, file=${params.folder}/${params.fname}. ${err.message}`, err)
+                log.error(`🌐🛑[uibuilder:apiv2:uibputfile] Admin API. File write FAIL. url=${params.url}, file=${params.folder}/${params.fname}. ${err.message}`, err)
                 res.statusMessage = err.message
                 res.status(500).end()
             } else {
                 // Send back a response message and code 200 = OK, 500 (Internal Server Error)=Update failed
-                log.trace(`✔️ [uibuilder:apiv2:uibputfile] Admin API. File write SUCCESS. url=${params.url}, file=${params.folder}/${params.fname}`)
+                log.trace(`🌐[uibuilder:apiv2:uibputfile] Admin API. File write SUCCESS. url=${params.url}, file=${params.folder}/${params.fname}`)
                 res.statusMessage = 'File written successfully'
                 res.status(200).end()
                 // Reload connected clients if required by sending them a reload msg
@@ -590,7 +589,7 @@ function adminRouterV2(uib, log) {
      * Also allows confirmation of whether a url is in use ('check' parameter) or a simple list of urls in use.
      */
     v2AdminRouter.get('/uibindex', function(/** @type {express.Request} */ req, /** @type {express.Response} */ res) {
-        log.trace('[uibindex] User Page/API. List all available uibuilder endpoints')
+        log.trace('🌐[uibuilder:apiv2:uibindex] User Page/API. List all available uibuilder endpoints')
 
         // If using own Express server, correct the URL's
         const url = new URL(req.headers.referer)
@@ -643,7 +642,7 @@ function adminRouterV2(uib, log) {
 
         // Validate the npm command to be used.
         if ( params.cmd === undefined ) {
-            log.error('[uibuilder:apiv2:uibnpmmanage] uibuilder Admin API. No command provided for npm management.')
+            log.error('🌐🛑[uibuilder:apiv2:uibnpmmanage] uibuilder Admin API. No command provided for npm management.')
             res.statusMessage = 'npm command parameter not provided'
             res.status(500).end()
             return
@@ -655,7 +654,7 @@ function adminRouterV2(uib, log) {
                 break
 
             default:
-                log.error('[uibuilder:apiv2:uibnpmmanage] Admin API. Invalid command provided for npm management.')
+                log.error('🌐🛑[uibuilder:apiv2:uibnpmmanage] Admin API. Invalid command provided for npm management.')
                 res.statusMessage = 'npm command parameter is invalid'
                 res.status(500).end()
                 return
@@ -664,13 +663,13 @@ function adminRouterV2(uib, log) {
         // package name must not exceed 255 characters
         // we have to have a package name
         if ( params.package === undefined ) {
-            log.error('[uibuilder:apiv2:uibnpmmanage] Admin API. package parameter not provided')
+            log.error('🌐🛑[uibuilder:apiv2:uibnpmmanage] Admin API. package parameter not provided')
             res.statusMessage = 'package parameter not provided'
             res.status(500).end()
             return
         }
         if ( params.package.length > 255 ) {
-            log.error('[uibuilder:apiv2:uibnpmmanage] Admin API. package name parameter is too long (>255 characters)')
+            log.error('🌐🛑[uibuilder:apiv2:uibnpmmanage] Admin API. package name parameter is too long (>255 characters)')
             res.statusMessage = 'package name parameter is too long. Max 255 characters'
             res.status(500).end()
             return
@@ -681,7 +680,7 @@ function adminRouterV2(uib, log) {
             // @ts-ignore
             const chkUrl = chkParamUrl(params)
             if ( chkUrl.status !== 0 ) {
-                log.error(`[uibuilder:apiv2:uibnpmmanage] Admin API. ${chkUrl.statusMessage}`)
+                log.error(`🌐🛑[uibuilder:apiv2:uibnpmmanage] Admin API. ${chkUrl.statusMessage}`)
                 res.statusMessage = chkUrl.statusMessage
                 res.status(chkUrl.status).end()
                 return
@@ -692,7 +691,7 @@ function adminRouterV2(uib, log) {
 
         const folder = RED.settings.userDir
 
-        log.info(`[uibuilder:apiv2:uibnpmmanage] Admin API. Running npm ${params.cmd} for package ${params.package} with tag/version '${params.tag ? params.tag : ''}'`)
+        log.info(`🌐📘[uibuilder:apiv2:uibnpmmanage] Admin API. Running npm ${params.cmd} for package ${params.package} with tag/version '${params.tag ? params.tag : ''}'`)
 
         // delete package lock file as it seems to mess up sometimes - no error if it fails
         fslib.removeSync(path.join(folder, 'package-lock.json'))
@@ -721,7 +720,7 @@ function adminRouterV2(uib, log) {
                     })
                     .catch((err) => {
                         // err has extra props: {all:string, code:number, command:string}
-                        log.error(`🛑 [uibuilder:apiv2:uibnpmmanage:${params.cmd}] Admin API. ERROR Running: \n'${err.command}' \n${err.all}`)
+                        log.error(`🌐🛑[uibuilder:apiv2:uibnpmmanage:${params.cmd}] Admin API. ERROR Running: \n'${err.command}' \n${err.all}`)
                         res.json({ 'success': false, 'result': err.all })
                         return false
                     })
@@ -745,15 +744,14 @@ function adminRouterV2(uib, log) {
                         return true
                     })
                     .catch((err) => {
-                        // log.warn(`[uibuilder:apiv2:uibnpmmanage] Admin API. ERROR Running npm ${params.cmd} for package ${params.package}`, err.stdout)
-                        log.warn(`⚠️ [uibuilder:apiv2:uibnpmmanage:remove] Admin API. ERROR Running: \n'${err.command}' \n${err.all}`)
+                        log.warn(`🌐⚠️[uibuilder:apiv2:uibnpmmanage:remove] Admin API. ERROR Running: \n'${err.command}' \n${err.all}`)
                         res.json({ 'success': false, 'result': err.all })
                         return false
                     })
                 break
             }
             default: {
-                log.error(`🛑 [uibuilder:apiv2:uibnpmmanage] Admin API. Command ${params.cmd} is not a valid command. Must be 'install', 'remove' or 'update'.`)
+                log.error(`🌐🛑[uibuilder:apiv2:uibnpmmanage] Admin API. Command ${params.cmd} is not a valid command. Must be 'install', 'remove' or 'update'.`)
                 res.statusMessage = 'No valid npm command available'
                 res.status(500).end()
                 break
